consolidate params
