--- core/dbt/cli/params/parse.py
@@ -1,13 +1,0 @@
-import click
-
-compile = click.option(
-    "--compile/--no-compile",
-    help="TODO: No help text currently available",
-    default=True,
-)
-
-write_manifest = click.option(
-    "--write-manifest/--no-write-manifest",
-    help="TODO: No help text currently available",
-    default=True,
-)
